+++ src/Addons/34thPRC_Overrides/data/nvgs/config_weapons.hpp
@@ -1,0 +1,6 @@
+#include "mvi\config_weapons.hpp"
+#include "mvi_cnm\config_weapons.hpp"
+#include "mvi_hul\config_weapons.hpp"
+#include "mvi_ul\config_weapons.hpp"
+#include "mvi_ul_cnm\config_weapons.hpp"
+#include "mvi_ul_hul\config_weapons.hpp"