+++ src/Addons/34thPRC_Overrides/data/nvgs/mvi/config_weapons.hpp
@@ -1,0 +1,11 @@
+class OPTRE_NVG;
+class OPTRE_NVG_MVI: OPTRE_NVG
+{
+	visionMode[]=
+		{
+			"Normal",
+			"NVG",
+			"TI"
+		};
+	thermalMode[]={0, 1};
+};