+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_standard/marines2/config.hpp
@@ -175,7 +175,7 @@ class 34thPRC_ArmourStandard_EOD_Helmet_Urban
 		label="Helmet Type";
 		alwaysSelectable=1;
 		changeingame=0;
-		values[]={"eod", "gungnir", "jfo"};
+		values[]={"eod", "gungnir", "jfo", "cqc", "commando", "grenadier", "scout"};
 		class eod
 		{
 			label="EOD";
@@ -187,6 +187,22 @@ class 34thPRC_ArmourStandard_EOD_Helmet_Urban
 		class jfo
 		{
 			label="JFO"
+		};
+		class cqc
+		{
+			label="CQC";
+		};
+		class commando
+		{
+			label="Commando";
+		};
+		class grenadier
+		{
+			label="Grenadier";
+		};
+		class scout
+		{
+			label="Scout";
 		};
 	};
 	class camo