+++ src/Addons/34thPRC_ArmourStandard/data/marines2/specialist_helmets/config_weapons.hpp
@@ -1,3 +1,7 @@
 #include "EOD\config_weapons.hpp"
 #include "gungnir\config_weapons.hpp"
 #include "JFO\config_weapons.hpp"
+#include "CQC\config_weapons.hpp"
+#include "commando\config_weapons.hpp"
+#include "grenadier\config_weapons.hpp"
+#include "scout\config_weapons.hpp"