+++ src/Addons/34thPRC_Overrides/data/nvgs/mvi_ul_cnm/config_weapons.hpp
@@ -1,0 +1,10 @@
+class OPTRE_NVG_MVI_UL_CNM: OPTRE_NVG
+{
+	visionMode[]=
+		{
+			"Normal",
+			"NVG",
+			"TI"
+		};
+	thermalMode[]={0, 1};
+};