+++ src/changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Scout for marines /ODSTs
 ### Updated
 - All Spartan shoulders have gunbags
+- Added black hot and white hot thermal modes to UNSC Marksma's Visual Interface and all MVI tagged NVGs
 
 ## 0.20.0
 ### Added