+++ src/changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Grenadier for Marines/ODSTs
 - Scout for marines /ODSTs
+- Shewolf Custom Helmet
 ### Updated
 - All Spartan shoulders have gunbags
 - Added black hot and white hot thermal modes to UNSC Marksma's Visual Interface and all MVI tagged NVGs